requirements.txt
pandas
requests
pyarrow
unidecode
apache-airflow==2.7.2
pytest